voice/requirements.txt
# Voice Pipeline Dependencies for Jetson Nano

# Speech-to-text (CUDA-optimized Whisper)
faster-whisper>=1.0.0

# Wake word detection
openwakeword>=0.6.0

# Audio processing
pyaudio>=0.2.14
numpy>=1.24.0,<2
sounddevice>=0.4.6

# Voice activity detection (pre-built aarch64 wheels)
webrtcvad-wheels>=2.0.10

# Anthropic API for intent resolution
anthropic>=0.26.0

# WebSocket for display communication
websockets>=12.0

# Configuration
pyyaml>=6.0

# Async utilities
aiohttp>=3.9.0
